extract network into its own class
